--- crates/quspin-core/src/basis/mod.rs
@@ -1,2 +1,0 @@
-// Forwarding shim — basis types now live in `quspin-basis`.
-pub use quspin_basis::*;
