--- crates/quspin-core/src/bitbasis/mod.rs
@@ -1,2 +1,0 @@
-// Forwarding shim — bitbasis types now live in `quspin-bitbasis`.
-pub use quspin_bitbasis::*;
