--- crates/quspin-core/src/expm/compute.rs
@@ -1,2 +1,0 @@
-// Forwarding shim — see `quspin-types` crate.
-pub use quspin_types::compute::*;
