--- crates/quspin-core/src/operator/bond/mod.rs
@@ -1,5 +1,0 @@
-pub mod dispatch;
-
-pub use dispatch::BondOperatorInner;
-pub use quspin_operator::bond::operator;
-pub use quspin_operator::bond::{BondOperator, BondTerm};
