--- crates/quspin-core/src/operator/boson/mod.rs
@@ -1,5 +1,0 @@
-pub mod dispatch;
-
-pub use dispatch::BosonOperatorInner;
-pub use quspin_operator::boson::operator;
-pub use quspin_operator::boson::{BosonOp, BosonOpEntry, BosonOperator};
