--- crates/quspin-core/src/operator/monomial/mod.rs
@@ -1,5 +1,0 @@
-pub mod dispatch;
-
-pub use dispatch::MonomialOperatorInner;
-pub use quspin_operator::monomial::operator;
-pub use quspin_operator::monomial::{MonomialOperator, MonomialTerm};
