--- crates/quspin-core/src/operator/pauli/mod.rs
@@ -1,5 +1,0 @@
-pub mod dispatch;
-
-pub use dispatch::HardcoreOperatorInner;
-pub use quspin_operator::pauli::operator;
-pub use quspin_operator::pauli::{HardcoreOp, HardcoreOperator, OpEntry};
