+++ crates/quspin-krylov/Cargo.toml
@@ -1,0 +1,11 @@
+[package]
+name = "quspin-krylov"
+version.workspace = true
+edition.workspace = true
+
+[dependencies]
+quspin-types = { path = "../quspin-types" }
+num-complex = { workspace = true }
+ndarray = { workspace = true }
+nalgebra = { workspace = true }
+rayon = { workspace = true }
