+++ crates/quspin-krylov/src/eig.rs
