+++ crates/quspin-krylov/src/ftlm_dynamic.rs
