+++ crates/quspin-krylov/src/lib.rs
