+++ crates/quspin-matrix/src/hamiltonian/inner.rs
