+++ crates/quspin-matrix/src/hamiltonian/mod.rs
