+++ crates/quspin-matrix/src/qmatrix/ops.rs
